linear and cubic Interpolation
Added code so that linear and cubic interpolation are possible.
Actually, I allowed scipy.interpolate.interp1d to be used, but I am not totally sure yet how powerful this function is.
Comparing linear with cubic interpolation, I think that cubic might be better. But more testing is needed.
Also comparing the interpolations with different kind of averaging is needed.